requirements.txt
#Remove the packages you don't need and install the rest using pip install -r requirements.txt

# Core Python and Django packages
django>=4.0  # Ensure compatibility with Python 3.9+
djangorestframework>=3.15.2  # For REST API development
django-extensions>=3.2.3  # Useful extensions for Django

# REST API and serialization
djangorestframework-simplejwt>=5.2.2  # For JWT-based authentication
django-cors-headers>=4.0.0  # CORS handling for APIs

# Natural Language Processing (AI)
transformers>=4.31.0  # Hugging Face Transformers for AI models
torch>=2.0.0  # PyTorch for running AI models
sentencepiece>=0.1.99  # Tokenizer library required by some models
sentence-transformers>=2.2.2  # Sentence Transformers for sentence embeddings

# Database and ORM
sqlalchemy>=1.4,<2.0  # For database management (compatible with Django and FastAPI)
psycopg2>=2.9.6  # PostgreSQL adapter for Python (replace with your database adapter if needed)

# Logging and debugging tools
loguru>=0.6.0  # Easy-to-use logging package
django-debug-toolbar>=4.0  # Django debugging tool
sentry-sdk>=1.29.2  # Sentry for error monitoring

# Performance optimization
Cython>=3.0.11  # Optional, for performance improvements
uvloop>=0.17.0  # High-performance asyncio loop for FastAPI

# Environment management
python-dotenv>=1.0.0  # Environment variable management

# HTTP utilities
httplib2>=0.22.0  # HTTP client
httptools>=0.6.1  # HTTP parsing (used by Uvicorn)
requests>=2.31.0  # Popular HTTP client library

# WebSocket support
channels>=4.0.0  # WebSockets for Django
channels-redis>=4.0.0  # Redis backend for Django channels

# Testing and development
pytest>=7.4.0  # Python testing framework
pytest-django>=4.5.2  # Pytest support for Django
factory-boy>=3.3.0  # Fixtures for testing  
flake8>=4.0.1  # Code linter

# Optional: API documentation
mkdocs>=1.5.2  # Static site generator for API documentation
mkdocs-material>=9.0.5  # Material design theme for MkDocs

# Data Science and Machine Learning
scikit-learn>=1.0.2  # Machine learning library
pandas>=1.3.5  # Data manipulation and analysis
numpy>=1.21.4  # Numerical computing library
